+++ packages/ng-helmet/src/lib/ng-helmet.module.ts
@@ -1,7 +1,7 @@
 import { ModuleWithProviders, NgModule } from "@angular/core";
 import { NgHelmetComponent } from "./ng-helmet.component";
-import { NgHelmetConfig } from "./ng-helmet.config";
-import { NG_HELMET_CONFIG_TOKEN } from "./ng-helmet.tokens";
+import { NgHelmetOptions } from "./ng-helmet.options";
+import { NG_HELMET_OPTIONS } from "./ng-helmet.tokens";
 import { NgHelmetChildDirective } from "./ng-helmet-child.directive";
 
 const components = [NgHelmetComponent, NgHelmetChildDirective];
@@ -11,10 +11,10 @@ const components = [NgHelmetComponent, NgHelmetChildDirective];
   exports: components,
 })
 export class NgHelmetModule {
-  static forRoot(config: NgHelmetConfig): ModuleWithProviders<NgHelmetModule> {
+  static forRoot(config: NgHelmetOptions): ModuleWithProviders<NgHelmetModule> {
     return {
       ngModule: NgHelmetModule,
-      providers: [{ provide: NG_HELMET_CONFIG_TOKEN, useValue: config }],
+      providers: [{ provide: NG_HELMET_OPTIONS, useValue: config }],
     };
   }
 }